+++ packages/ente-ui/src/elements/PasswordRequirementsHint.tsx
@@ -6,7 +6,7 @@ const en: React.SFC = () => {
       <p>Requirements:</p>
       <ul>
         <li>8 to 100 characters</li>
-        <li>at least 1 upper-case letter</li>
+        <li>at least 1 number</li>
         <li
         >{`at least 1 special character ([!@#$%^&*()_+\-=\[\]{};':"\\|,<>\/?])`}</li>
       </ul>
@@ -20,14 +20,14 @@ const de: React.SFC = () => {
       Vorraussetzungen:
       <ul>
         <li>8 bis 100 Zeichen</li>
-        <li>Mindestens 1 Großbuchstabe</li>
+        <li>Mindestens 1 Zahl</li>
         <li
         >{`Mindestens 1 Sonderzeichen ([!@#$%^&*()_+\-=\[\]{};':"\\|,<>\/?])`}</li>
       </ul>
     </>
   );
 };
-export const PasswordRequirements = {
+export const PasswordRequirementsHint = {
   en,
   de
 };